+++ Sources/FBSDKCoreKit_Basics/include/FBSDKCrashHandler.h
@@ -18,8 +18,8 @@ NS_SWIFT_NAME(CrashHandler)
 
 @property (class, nonatomic, readonly) FBSDKCrashHandler *shared;
 
-- (instancetype)init DEPRECATED_MSG_ATTRIBUTE("`init` is deprecated and will be removed in the next major release.");
-+ (instancetype)new DEPRECATED_MSG_ATTRIBUTE("`new` is deprecated and will be removed in the next major release.");
+- (instancetype)init NS_UNAVAILABLE;
++ (instancetype)new NS_UNAVAILABLE;
 
 + (void)disable;
 + (void)addObserver:(id<FBSDKCrashObserving>)observer;